infra/requirements.txt
aws-cdk-lib==2.158.0
constructs>=10.0.0,<11.0.0
python-dotenv>=0.15.0,<1.0.0